# openapi-spec.yaml
openapi: 3.0.3
info:
  title: API de Clientes e Pagamentos
  description: API para consulta de clientes e estatísticas de pagamentos.
  version: 1.0.0
servers:
  - url: http://localhost:80
paths:
...
  /clientes/{id}:
    get:
      summary: Busca cliente por ID
      operationId: buscaCliente
      tags:
        - Clientes
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único do cliente.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Dados do cliente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Cliente não encontrado.
components:
  schemas:
    Cliente:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 123
        customer_name:
          type: string
          example: João da Silva
...